:adhesive_bandage: (sync_fns): Use checksums for change detection
Although utilising checksums is slower than relying on file size and last modified timestamps for change detection, in some situations it is more reliable. For example, in instances such as transferring files between different file systems which record timestamps with differing levels of precision. Not using checksums in this instance would mean that rsync would rightly interpret the change in precision as a timestamp change, although this might be misleading for a user who is unaware of the nuances between different file systems. Whilst it is possible that we could make this configurable in the future for bigger repos where speed might be an issue, exposing too many options for the user at this stage goes against the grain of keeping the utility simple to use by employing sensible defaults.