💥 (glit): Change bit values to true or false cmds (#19)
* :boom: (glit): Change bit values to true or false cmds

Includes breaking changes. This commit replaces bit values of `1` and `0` with `true` and `false` bash commands. Prior to this change, we were using a value of `1` to represent a 'truthy' value for use in conditional statements, instead we now utilise the shell command `true` which actually produces a status code of `0` indicating 'success'. For developers new to shell scripting or coming from languages where a bit value of `1` would typically represent a 'truthy' value, this may seem unintuitive at first, however a zero status code to indicate 'success' and a non-zero status code to indicate 'failure' is more idiomatic to shell scripting. For readability purposes, instead of using status codes as numeric values, we use the `true` and `false` commands which return `0` and `1` respectively.

* :bookmark: (variables): Bump minor version

The changes associated with this version bump include breaking changes. Whilst it would be typical practice to increment a major version due to breaking changes, given the application is not yet at version 1.0.0, a minor version increment is acceptable according to the Semantic Versioning 2.0.0 spec.